[ot] hw/core: Use resettable API to maintain CPU on reset
Add the capability to maintain CPU in reset.

The Resettable API has added a new way to manage reset, where resetting a device
not only trigger a transient reset, but enable devices to be held in reset and
manage a reset tree.

This commit adds this resettable feature to CPUs, so they can be held in reset
and released once other devices in the machine are released. The CPU reset is
no longer a transient state but a stable state. The CPU reset state can be
tracked with a new attribute: held_in_reset.

This feature is useful to better emulate SoC like OpenTitan where the CPU is
actively held in Reset while other devices complete their own initialization,
such as the power manager and ROM controller.

It also enables to update the CPU configuration and attributes while the CPUs
are held in reset. On RISC-V CPUs, reset vector and mtvec registers for example
can be updated before releasing the harts from reset.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>