[ot] hw/opentitan: ot_rstmgr: fix a bug on vCPU reset
It is mandatory to wait for the vCPU to actually stop before resuming
the reset sequence. Failing to do so may lead TCG to execute a cached
TB from the new PC, updated from reset vector.

Avoid using pause_all_vcpus/resume_all_vcpus as other SoCs in the
machine should not be stopped. Moreover the virtual clock should not be
halted as other SoCs may keep running while the local reset sequence is
ran.

Avoid using qemu_system_reset_request for the same reason, which is
even more intrusive for the whole VM.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>