[ot] jtag: rework JTAG / TAP controller support.
Initial JTAG implementation had borrowed GDB server to support incoming connections.

It is better to use a regular QEMU device: it can be instantiated multiple times
and does not need to rely on global variables as GDB server does.

JTAG / TAP controller service is enabled when a CharDev is connected to the TAP
controller instance. This CharDev may be created from the command line. There is
no longer any need for a special `-jtag` option in `vl.c`

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>