[ot] scripts/opentitan: pyot.py: add a text-based trigger matching feature.
When a `with` context is defined. i.e. one or more apps that should execute
once the VM is up and running and usually communicate with the VM over a
dedicated channel, it is now possible to postpone the context execution
when a specific string is matched on the VM default virtual comm port.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>