[ot] hw/opentitan: ot_timer: do not reconfigure timer when max time is reached
INT64_MAX is mostly computed when the guess OS perform a read-modify-write
sequence of the 64-bit comparator with successive 32-bit register access.
It first writes max time to timer hi register so that updating timer lo does
not trigger a spurious timer alarm/timeout, then update a second time timer hi
with the real value.
This sequences leads to reconfigure the QEMU timer three times in a row, the two
first configs only leads to reconfigure the timer with a near-infinite time.

Signed-off-by: Emmanuel Blot <eblot@rivosinc.com>