[ot] hw/opentitan: ot_hmac: Implement key/digest size regs & checks
This commit implements initial register interface support for the key
length and digest size fields in the config register of OpenTitan's
HMAC. It checks that the values written are valid one-hot encodings, and
maps them as is done in current hardware if not.

This also implements some additional checks that the combination of (key
length, digest size, HMAC enable) is valid when attempting to run the
START/CONTINUE command, otherwise producing a (new type of) software
error.

Also fixes some comment formatting (code style).