[ot] hw/opentitan: ot_aes: preserve external state when a new message is started
The AES block does not clear the CPU-visible registers when starting a
new message. Only "the state registers inside the cipher core are cleared
with PRD during the last round of every encryption/decryption."

Fixing this allows the OpenTitan `aes_test.c` test to pass, as more recent
versions of it check the IV register state at the end of the test.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>