[ot] hw/opentitan: ot_aes: update IV registers after operation in CFB mode
This fix fixes the last known OT AES test failure (aes_functest) in our
current test set. The streaming encryption and decryption tests in CFB
mode would fail due to the IV registers not being properly updated (in
the aes_functest streaming mode we start and finish an operation for
each block).

The need for this fix suggests we might need an overall more systematic
approach to handling IV state. The current design should be revisited
at a later date.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>